wireguard: Send Keepalive on Error
According to the WireGuard white paper, keepalive packets should be
sent, if a message was received and no message was sent back after
Keepalive-Timeout seconds (10 s). In salty-stun however, if there's an
error in any of the upper layers, a message will never be sent back,
because salty-stun only sends messages as a response to a request. That
means we don't have to wait for the timeout, but can immediately send a
keepalive message back. By doing that, the initiator (i.e. client)
knows, that the connection is still alive.