Add custom hashtable implementation
The cdsa hashtable doesn't work well for incrementing a counter in an
entry, because you have to first lookup the entry and if it doesn't
exist, add it. With cdsa this would mean calculating the hash twice.
Incrementing a counter is a use case for rate limiting per source IP.

Add a simpler implementation with a separated bucket index calculation.
The index can then be reused for lookup and insertion.